feat(gas): implement EIP-8037 issue #2 — 0→x→0 storage reservoir refill
When a storage slot is restored to its original zero value within the
same transaction, the state gas originally charged for the 0→x
transition is now returned directly to the reservoir rather than
routed through the capped (spent/5) refund counter. This preserves
the EIP-8037 invariant that state gas consumption correlates with
actual state creation.

Changes:
- `GasTracker.state_gas_spent` becomes `i64`. A child frame's count
  can legitimately go negative when it clears a slot that the parent
  had set; the net is reconciled on frame return.
- New `GasTracker::refill_reservoir(amount)` and `Gas::refill_reservoir`
  wrappers: add to reservoir, subtract from `state_gas_spent`.
- New `GasParams::sstore_state_gas_refill` returns `32 * CPSB` only
  on 0→x→0 restoration.
- SSTORE instruction calls `refill_reservoir` when EIP-8037 is
  enabled and the restoration condition matches.
- EIP-8037 table: `sstore_set_refund` drops from `32 * CPSB + 2800`
  to `2800`; the state portion now flows through the reservoir refill
  path rather than the refund counter.
- `handle_reservoir_remaining_gas` (frame.rs) uses saturating i64
  math on both success and revert/halt paths so the parent's matching
  0→x charge is correctly netted out on success, and the combined
  `state_gas_spent + reservoir` is clamped to 0 on failure.
- Top-frame reconciliation (handler.rs) and `build_result_gas`
  (post_execution.rs) clamp to 0 before the public `u64` surface.
- Precompile + inspector plumbing updated for the i64 switch.

Test updated: `test_eip8037_sstore_set_then_clear_refund` now
asserts `state_gas_spent == 0` and that total gas spent matches the
baseline (was asserting +200k and a strict `>` inequality). Golden
JSON regenerated.

+++ crates/handler/src/precompile_provider.rs
@@ -97,7 +97,7 @@ pub fn precompile_output_to_interpreter_result(
     };
 
     // set state gas, reservoir is already set in the Gas constructor
-    result.gas.set_state_gas_spent(output.state_gas_used);
+    result.gas.set_state_gas_spent(output.state_gas_used as i64);
     result.gas.record_refund(output.gas_refunded);
 
     // spend used gas.
